refactor(dashboard): drop Institution episode surfaces from the TS dashboard
Companion to the server-side Institution removal. The memory-subsystems
page keeps hebbian synapses, recall quality, and delegation requests;
everything that rendered the dead institution_episodes.jsonl store is
gone: the episodes section and EpisodeCard, the 'episodes' focus value
(the focus concept had no other value left), the keeper/outcome/q
filter bar, the synapse-pair-to-episode filter hop (the matrix itself
stays), the MemorySubsystemsEpisode API type and its fetch params, and
the architecture-flow diagram node. agent-detail-memory and lab
embeddings updated alongside their tests; the focus test file is
deleted with the concept.

No local typecheck (no node_modules in this worktree); dashboard CI is
the compile gate.
