+++ dashboard/src/config/navigation.ts
@@ -459,7 +459,7 @@ export const DASHBOARD_SECTION_ITEMS: Record<NonHomeTabId, DashboardSectionNavIt
     {
       id: 'memory-subsystems',
       label: 'Memory OS',
-      description: 'Live episodes, user model projection, Hebbian synapses, and gated memory entries.',
+      description: 'Hebbian synapses, recall quality, and delegation requests.',
       params: { section: 'memory-subsystems' },
     },
     {
